backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
mysql-connector-python==8.2.0  # MySQL connector
pymysql==1.1.0  # Alternative MySQL driver
# psycopg2-binary==2.9.9  # Only for PostgreSQL
# aiosqlite==0.19.0  # For SQLite (deprecated)

# Redis and Celery
redis==5.0.1
celery==5.3.4

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# AI and ML
openai==1.3.7
Pillow==10.1.0
librosa==0.10.1
numpy==1.26.2

# WebSocket
websockets==12.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Monitoring and logging
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx-test==0.24.0
